+++ tsconfig.paths.json
@@ -2,72 +2,32 @@
   "compilerOptions": {
     "baseUrl": "./src",
     "paths": {
-      "@src": [
-        "./"
-      ],
-      "@src/*": [
-        "./*"
-      ],
-      "@apis": [
-        "./apis"
-      ],
-      "@apis/*": [
-        "./apis/*"
-      ],
-      "@assets": [
-        "./assets"
-      ],
-      "@assets/*": [
-        "./assets/*"
-      ],
-      "@components": [
-        "./components"
-      ],
-      "@components/*": [
-        "./components/*"
-      ],
-      "@containers": [
-        "./containers"
-      ],
-      "@containers/*": [
-        "./containers/*"
-      ],
-      "@hooks": [
-        "./hooks"
-      ],
-      "@hooks/*": [
-        "./hooks/*"
-      ],
-      "@interfaces": [
-        "./interfaces"
-      ],
-      "@interfaces/*": [
-        "./interfaces/*"
-      ],
-      "@router": [
-        "./router"
-      ],
-      "@router/*": [
-        "./router/*"
-      ],
-      "@store": [
-        "./store"
-      ],
-      "@store/*": [
-        "./store/*"
-      ],
-      "@styles": [
-        "./styles"
-      ],
-      "@styles/*": [
-        "./styles/*"
-      ],
-      "@pages": [
-        "./pages"
-      ],
-      "@pages/*": [
-        "./pages/*"
-      ]
+      "@src": ["./"],
+      "@src/*": ["./*"],
+      "@apis": ["./apis"],
+      "@apis/*": ["./apis/*"],
+      "@assets": ["./assets"],
+      "@assets/*": ["./assets/*"],
+      "@components": ["./components"],
+      "@components/*": ["./components/*"],
+      "@containers": ["./containers"],
+      "@containers/*": ["./containers/*"],
+      "@hooks": ["./hooks"],
+      "@hooks/*": ["./hooks/*"],
+      "@interfaces": ["./interfaces"],
+      "@interfaces/*": ["./interfaces/*"],
+      "@router": ["./router"],
+      "@router/*": ["./router/*"],
+      "@store": ["./store"],
+      "@store/*": ["./store/*"],
+      "@styles": ["./styles"],
+      "@styles/*": ["./styles/*"],
+      "@pages": ["./pages"],
+      "@pages/*": ["./pages/*"],
+      "@utils": ["./utils"],
+      "@utils/*": ["./utils/*"],
+      "@constants": ["./constants"],
+      "@constants/*": ["./constants/*"]
     }
   }
 }
